backend/requirements.txt
# ─── BhumiRaksha Backend Dependencies ─────────────────────────────────

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
geoalchemy2==0.15.2

# Redis
redis==5.2.1
aioredis==2.0.1

# ML / Computer Vision
ultralytics==8.3.52
torch==2.5.1
torchvision==0.20.1
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==2.2.1
transformers==4.47.1

# Geospatial
earthengine-api==1.4.3
geopandas==1.0.1
shapely==2.0.6
rasterio==1.4.3
pyproj==3.7.0

# Image Processing
exifread==3.0.0
imagehash==4.3.1

# HTTP / API
httpx==0.28.1
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# SMS
requests==2.32.3

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Dev
black==24.10.0
ruff==0.8.6
